Первичный рефакторинг утилит
