+++ src/const.js
@@ -1,5 +1,7 @@
 // Исходные данные для всего приложения, межкомпонентные
 
+export const NUMERAL_SYSTEM_BASE = 10;
+
 export const FILTERS = [`everything`, `future`, `past`];
 
 export const SORTINGS = [`day`, `event`, `time`, `price`, `offer`];
